+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/VisualBasic/Extensions/SyntaxTreeExtensions.vb
@@ -5,6 +5,7 @@
 Imports System.Runtime.CompilerServices
 Imports System.Runtime.InteropServices
 Imports System.Threading
+Imports Microsoft.CodeAnalysis.VisualBasic.Extensions.ContextQuery
 Imports Microsoft.CodeAnalysis.VisualBasic.Syntax
 
 Namespace Microsoft.CodeAnalysis.VisualBasic.Extensions
@@ -325,8 +326,17 @@ recurse:
 
         <Extension()>
         Public Function IsInPreprocessorDirectiveContext(syntaxTree As SyntaxTree, position As Integer, cancellationToken As CancellationToken) As Boolean
+            Dim directive As DirectiveTriviaSyntax = Nothing
+            Return IsInPreprocessorDirectiveContext(syntaxTree, position, cancellationToken, directive)
+        End Function
+
+        Friend Function IsInPreprocessorDirectiveContext(
+                syntaxTree As SyntaxTree,
+                position As Integer,
+                cancellationToken As CancellationToken,
+                ByRef directive As DirectiveTriviaSyntax) As Boolean
             Dim token = syntaxTree.FindTokenOnLeftOfPosition(position, cancellationToken, includeDirectives:=True, includeDocumentationComments:=True)
-            Dim directive = token.GetAncestor(Of DirectiveTriviaSyntax)()
+            directive = token.GetAncestor(Of DirectiveTriviaSyntax)()
 
             ' Directives contain the EOL, so if the position is within the full span of the
             ' directive, then it is on that line, the only exception is if the directive is on the
@@ -337,9 +347,8 @@ recurse:
                 Return False
             End If
 
-            Return _
-                      directive.FullSpan.Contains(position) OrElse
-                      directive.FullSpan.End = syntaxTree.GetRoot(cancellationToken).FullSpan.End
+            Return directive.FullSpan.Contains(position) OrElse
+                   directive.FullSpan.End = syntaxTree.GetRoot(cancellationToken).FullSpan.End
         End Function
 
         <Extension()>
