+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/VisualBasic/Extensions/ContextQuery/VisualBasicSyntaxContext.vb
@@ -65,6 +65,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Extensions.ContextQuery
             isNamespaceContext As Boolean,
             isNamespaceDeclarationNameContext As Boolean,
             isPreProcessorDirectiveContext As Boolean,
+            isPreProcessorExpressionContext As Boolean,
             isRightOfNameSeparator As Boolean,
             isSingleLineStatementContext As Boolean,
             isExpressionContext As Boolean,
@@ -90,6 +91,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Extensions.ContextQuery
                 isNamespaceContext,
                 isNamespaceDeclarationNameContext,
                 isPreProcessorDirectiveContext,
+                isPreProcessorExpressionContext,
                 isRightOfNameSeparator,
                 isStatementContext:=isSingleLineStatementContext,
                 isAnyExpressionContext:=isExpressionContext,
@@ -155,6 +157,7 @@ Namespace Microsoft.CodeAnalysis.VisualBasic.Extensions.ContextQuery
                 isNamespaceContext:=syntaxTree.IsNamespaceContext(position, targetToken, cancellationToken, semanticModel),
                 isNamespaceDeclarationNameContext:=syntaxTree.IsNamespaceDeclarationNameContext(position, cancellationToken),
                 isPreProcessorDirectiveContext:=syntaxTree.IsInPreprocessorDirectiveContext(position, cancellationToken),
+                isPreProcessorExpressionContext:=syntaxTree.IsInPreProcessorExpressionContext(position, cancellationToken),
                 isRightOfNameSeparator:=syntaxTree.IsRightOfDot(position, cancellationToken),
                 isSingleLineStatementContext:=syntaxTree.IsSingleLineStatementContext(position, targetToken, cancellationToken),
                 isExpressionContext:=syntaxTree.IsExpressionContext(position, targetToken, cancellationToken, semanticModel),
